Refactoring to clean
