+++ package.json
@@ -2,9 +2,9 @@
   "name": "node-locations",
   "version": "1.0.0",
   "description": "NodeJS Location Inquiry",
-  "main": "index.js",
+  "main": "src/app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha test",
     "start": "nodemon --exec babel-node src/server"
   },
   "repository": {
@@ -12,7 +12,7 @@
     "url": "git+https://github.com/ramirlm/node-locations.git"
   },
   "author": "Ramir Mesquita",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/ramirlm/node-locations/issues"
   },
@@ -20,6 +20,7 @@
   "dependencies": {
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
+    "lodash": "^4.17.15",
     "mongodb": "^3.3.0"
   },
   "devDependencies": {
@@ -28,6 +29,8 @@
     "babel-loader": "^8.0.6",
     "babel-preset-env": "^1.7.0",
     "babel-preset-es2015": "^6.24.1",
+    "chai": "^4.2.0",
+    "mocha": "^6.2.0",
     "nodemon": "^1.19.1",
     "webpack": "^4.39.2",
     "webpack-cli": "^3.3.6"
